# Coverage: Spawn in sub-process