--- a/code/addons/experimental-coverage/package.json
+++ b/code/addons/experimental-coverage/package.json
@@ -88,7 +88,8 @@
     ],
     "nodeEntries": [
       "./src/preset.ts",
-      "./src/node/coverage-reporter.ts"
+      "./src/node/coverage-reporter.ts",
+      "./src/node/vitest.ts"
     ]
   },
   "gitHead": "e6a7fd8a655c69780bc20b9749c2699e44beae16",