## What I did

- spawn `vitest` in separate process, communicate via channel over `IPC` transport

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:
- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

     - `bug`: Internal changes that fixes incorrect behavior.
     - `maintenance`: User-facing maintenance tasks.
     - `dependencies`: Upgrading (sometimes downgrading) dependencies.
     - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
     - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
     - `documentation`: Documentation **only** changes. Will not show up in release changelog.
     - `feature request`: Introducing a new feature.
     - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
     - `other`: Changes that don't fit in the above categories.
   
   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  76.3 MB | 76.3 MB | 0 B | -0.73 | 0% |
| initSize |  177 MB | 169 MB | -7.95 MB | **-1.61** | -4.7% |
| diffSize |  101 MB | 92.8 MB | -7.95 MB | **-1.6** | 🔰-8.6% |
| buildSize |  7.46 MB | 7.46 MB | 0 B | **2.39** | 0% |
| buildSbAddonsSize |  1.62 MB | 1.62 MB | 0 B | **3.15** | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  2.3 MB | 2.3 MB | 0 B | **-3.16** | 0% |
| buildSbPreviewSize |  351 kB | 351 kB | 0 B | **3.16** | 0% |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  4.46 MB | 4.46 MB | 0 B | **3.14** | 0% |
| buildPreviewSize |  3 MB | 3 MB | 0 B | 0.7 | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  6.9s | 6.8s | -129ms | **-1.25** | -1.9% |
| generateTime |  19.4s | 20.8s | 1.4s | 0.13 | 6.8% |
| initTime |  18.6s | 18.4s | -163ms | 0.35 | -0.9% |
| buildTime |  12.7s | 11.5s | -1s -179ms | -0.86 | -10.2% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  7.8s | 7.5s | -282ms | -0.98 | -3.7% |
| devManagerResponsive |  5.3s | 4.9s | -345ms | -0.75 | -6.9% |
| devManagerHeaderVisible |  743ms | 747ms | 4ms | **-1.26** | 0.5% |
| devManagerIndexVisible |  782ms | 780ms | -2ms | **-1.29** | -0.3% |
| devStoryVisibleUncached |  1.2s | 1.2s | 29ms | -0.63 | 2.3% |
| devStoryVisible |  781ms | 779ms | -2ms | **-1.44** | -0.3% |
| devAutodocsVisible |  680ms | 685ms | 5ms | -0.83 | 0.7% |
| devMDXVisible |  602ms | 716ms | 114ms | -0.72 | 15.9% |
| buildManagerHeaderVisible |  642ms | 746ms | 104ms | -0.13 | 13.9% |
| buildManagerIndexVisible |  647ms | 747ms | 100ms | -0.17 | 13.4% |
| buildStoryVisible |  681ms | 805ms | 124ms | 0.06 | 15.4% |
| buildAutodocsVisible |  668ms | 679ms | 11ms | -0.18 | 1.6% |
| buildMDXVisible |  594ms | 641ms | 47ms | -0.5 | 7.3% |

<!-- BENCHMARK_SECTION -->

<!-- greptile_comment -->

## Greptile Summary

This PR implements running Vitest in a separate process for the experimental coverage addon, improving isolation and potentially performance.

- Added `code/addons/experimental-coverage/src/node/vitest.ts` to handle IPC communication for coverage data
- Modified `code/addons/experimental-coverage/src/preset.ts` to spawn Vitest as a child process using `child_process.fork()`
- Updated `code/addons/experimental-coverage/src/node/coverage-manager.ts` to work with the new VitestManager setup
- Added new node entry './src/node/vitest.ts' in `code/addons/experimental-coverage/package.json`

<!-- /greptile_comment -->